requirements.txt
openai
pyaudio
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
torch
git+https://github.com/openai/whisper.git


annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
distro==1.9.0
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.97.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
SpeechRecognition==3.14.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
